requirements.txt
# Core Framework
fastapi==0.103.1
uvicorn==0.23.2
flask==2.3.3
jinja2==3.1.2
pydantic==2.3.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
asyncpg==0.28.0

# NLP
spacy==3.6.1
nltk==3.8.1
scikit-learn==1.3.0
transformers==4.33.1

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.12.0
scrapy==2.10.0
python-dotenv==1.0.0

# Cloud & Deployment
boto3==1.28.40
gunicorn==21.2.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1 